refactor: move types to types.tsx, constants to constants.tsx, add public api
