+++ src/components/restful/index.ts
@@ -1,0 +1,1 @@
+export { RestfulView } from './restful-view';
