+++ src/components/restful/select-language/index.ts
@@ -1,0 +1,1 @@
+export { SelectLanguage } from './select-language';
