+++ src/components/restful/select-method/index.ts
@@ -1,0 +1,1 @@
+export { SelectMethod } from './select-method';
