+++ src/components/restful/select-method/select-method.tsx
@@ -1,22 +1,13 @@
 'use client';
 
-import { Option, SelectElement } from '@/components/restful/select/select';
-
-const RESTFUL_METHODS = [
-  'GET',
-  'POST',
-  'PUT',
-  'PATCH',
-  'DELETE',
-  'HEAD',
-] as const;
+import { SelectElement } from '@/components/restful/select';
+import { RESTFUL_METHODS } from '@/constants/constants';
+import { METHODS, Option } from '@/types/types';
 
 const METHODS_OPTIONS: Option[] = RESTFUL_METHODS.map((method) => ({
   value: method,
   label: method,
 }));
-
-export type METHODS = (typeof RESTFUL_METHODS)[number];
 
 export function SelectMethod({ currentMethod }: { currentMethod: METHODS }) {
   return (
