+++ src/components/restful/select/index.ts
@@ -1,0 +1,1 @@
+export { SelectElement } from './select';
